added a few updates to the manual uploading/deleting of an entry

+++ app.py
@@ -30,27 +30,28 @@ def home_page():
 # route that handles buttons on the home page
 @app.route("/modify", methods=['GET', 'POST'])
 def modify():
+    people = Person.objects()
     if request.method == 'POST':
         if request.form['action'] == 'add':
-            fname = request.form['fname_add']
-            lname = request.form['lname_add']
+            first_name = request.form['fname_add']
+            last_name = request.form['lname_add']
             email = request.form['email_add']
-            if fname and lname and email:
-                add = Person()
-                add.first_name = fname
-                add.last_name = lname
-                add.save()
-            count = Person.objects.count()
-            return render_template('add_delete.html', count=count)
+            save_dict = {"first_name": first_name, "last_name": last_name, "email": email}
+            if first_name and last_name and email:
+                num_uploads = parser.save_single(save_dict)
+            output_str = "Successfully Uploaded " + str(num_uploads)
+            return render_template('results.html', count=people.count(), entries=people, msg=output_str)
         elif request.form['action'] == 'delete':
-            query = request.form['fname_delete']
-            to_delete = Person.objects(first_name=query)
-            to_delete.delete()
-            count = Person.objects.count()
-            return render_template('add_delete.html', count=count)
+            query = request.form['email_delete']
+            to_delete = people(email=query)
+            num_delete = 0
+            if to_delete:
+                to_delete.delete()
+                num_delete = 1
+            output_str = "Successfully Deleted " + str(num_delete)
+            return render_template('results.html', count=people.count(), entries=people, msg=output_str)
     elif request.method == 'GET':
-        count = Person.objects.count()
-        return render_template('add_delete.html', count=count)
+        return render_template('add_delete.html', count=people.count())
 
 @app.route('/participants')
 def participants():
@@ -60,7 +61,7 @@ def participants():
     action = request.args.get('action', '')
     if action:
         if action == 'uploaded':
-            parser.save_data()
+            parser.save_group()
             count = Person.objects.count()
             entries = Person.objects()
             return render_template('results.html', count=count, entries=entries, msg="Successful Upload")
